% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distransam.R
\name{distransam_series}
\alias{distransam_series}
\title{Randomly samples groups that have series data to get equal N's}
\usage{
distransam_series(x, group, site, ID, series)
}
\arguments{
\item{series}{}
}
\value{
dataframe
}
\description{
Randomly samples groups that have series data to get equal N's
}
\examples{
distransam_series(worm_data_time_series, 'strain', 'plate', 'ID', 'time')
}

