% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distransam-package.r, R/distransam.R
\docType{package}
\name{distransam}
\alias{distransam}
\alias{distransam-package}
\title{distransam: randomly samples groups to get equal N's}
\usage{
distransam(x, grouping_var, sample_var = NULL, max_N = NULL)
}
\arguments{
\item{x}{dataframe}

\item{grouping_var}{string}

\item{sample_var}{string (optional)}

\item{max_N}{integer (not yet implemented)}
}
\value{
dataframe
}
\description{
Takes a data frame, and returns one with equal N's for each group. For unequal sample sizes, it finds the minimum sample size from the smallest group, and randomly selects that number of samples from each group. A single sub-groups can be specified, as can a maximum N value. The returned object is a dataframe, with all the same original columns as the original data frame.

Takes a data frame, and returns one with equal N's for each group. For unequal sample sizes, it finds the minimum sample size from the smallest group, and randomly selects that number of samples from each group. A single sub-groups can be specified, as can a maximum N value. The returned object is a dataframe, with all the same original columns as the original data frame.
}
\examples{
library(gapminder)
distransam(gapminder[gapminder$year == 1992,], 'continent')
}

