% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distransam.R
\name{distransam_series}
\alias{distransam_series}
\title{Randomly samples groups that have series data to get equal N's}
\usage{
distransam_series(x, grouping_var, subgroup_var = NULL, series_var)
}
\arguments{
\item{x}{}

\item{grouping_var}{}

\item{series_var}{}
}
\value{
dataframe
}
\description{
Randomly samples groups that have series data to get equal N's
}
\examples{

}

